adapt module transparency dialogue
