+++ webclient/app/src/app/features/info/ModuleInfo.jsx
@@ -1,6 +1,6 @@
 import CancelIcon from "@mui/icons-material/Cancel";
 import CheckCircleIcon from "@mui/icons-material/CheckCircle";
-import {Card, CardContent, Divider, IconButton, Table, TableBody, TableCell, TableContainer, TableRow, Tooltip, Typography} from '@mui/material';
+import {Card, CardContent, Divider, IconButton, Link, Table, TableBody, TableCell, TableContainer, TableRow, Tooltip, Typography} from '@mui/material';
 import React, {useEffect, useMemo, useState} from "react";
 import {useTranslation} from "react-i18next";
 import DescriptionIcon from '@mui/icons-material/Description';
